backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
django-mptt==0.16.0
dj-database-url==2.3.0
openpyxl==3.1.5
PyJWT==2.11.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
google-generativeai==0.8.3
pdfplumber==0.11.4
